+++ server/mailet/mailets/src/main/java/org/apache/james/transport/mailets/Bounce.java
@@ -25,6 +25,7 @@ import java.util.List;
 import javax.mail.MessagingException;
 import javax.mail.internet.InternetAddress;
 
+import org.apache.james.transport.mailets.redirect.AbstractRedirect;
 import org.apache.mailet.Mail;
 import org.apache.mailet.MailAddress;
