+++ server/mailet/mailets/src/main/java/org/apache/james/transport/mailets/Redirect.java
@@ -26,6 +26,8 @@ import javax.mail.MessagingException;
 import javax.mail.internet.AddressException;
 import javax.mail.internet.InternetAddress;
 
+import org.apache.james.transport.mailets.redirect.AbstractRedirect;
+import org.apache.james.transport.mailets.redirect.TypeCode;
 import org.apache.mailet.Mail;
 import org.apache.mailet.MailAddress;
