+++ server/mailet/mailets/src/test/java/org/apache/james/transport/mailets/ForwardTest.java
@@ -29,6 +29,7 @@ import java.util.Collection;
 import javax.mail.MessagingException;
 
 import org.apache.james.dnsservice.api.DNSService;
+import org.apache.james.transport.mailets.redirect.TypeCode;
 import org.apache.mailet.MailAddress;
 import org.apache.mailet.base.MailAddressFixture;
 import org.apache.mailet.base.test.FakeMailContext;
