+++ server/mailet/mailets/src/test/java/org/apache/james/transport/mailets/ResendTest.java
@@ -28,6 +28,7 @@ import java.net.UnknownHostException;
 import javax.mail.MessagingException;
 
 import org.apache.james.dnsservice.api.DNSService;
+import org.apache.james.transport.mailets.Resend;
 import org.apache.mailet.MailAddress;
 import org.apache.mailet.base.test.FakeMailContext;
 import org.apache.mailet.base.test.FakeMailetConfig;
